requirements.txt
fastapi
uvicorn[standard]
openai
scikit-learn
python-dotenv
